Update index.html
Medical fibers are the monomers of biotextiles that could be either natural polymers or synthetic (manmade ones). The difference between synthetic and natural polymers is that synthetic ones can be synthesized and modified in a controlled manner to achieve a specific properties for a given application. Both have shown to be effective as suitable suture materials especially natural polymers such as silk and linen fibers which have been used for over 500 years and non-absorbable polymers such as nylon, polypropylene  and other monofilament sutures with absorbable properties. Playing an integral role in medicine today they are manufactured using various methods such as melt extrusion, wet/gel spinning, electrospinning, centrifugal spinning and hydrogen fiber spinning all of which are going to be discussed more thoroughly in this article.